Fixed issues
In the Tetris.py file it looks for the file arcade.ttf with lowercase letters at the end, the real name of the arcade font is arcade.TTF with capital letters so I renamed the file.

When looking for files in the directory it looks at an absolute path which is /Users/rajat/PycharmProjects/Tetris/file. I made it look in the same directory as the python file so you could move it and it would still work.